feat: rename segmentation tab to segments
