+++ assets/wizards/popups/views/index.js
@@ -1,4 +1,4 @@
 export { default as Campaigns } from './campaigns';
 export { default as Analytics } from './analytics';
 export { default as Settings } from './settings';
-export { default as Segmentation } from './segmentation';
+export { default as Segments } from './segments';
